refactor to include Dataset interface in models
